Move ProductListRecyclerAdapter class
