+++ app/src/main/java/com/ratajczykdev/inventoryapp/ProductEditActivity.java
@@ -15,7 +15,6 @@ import android.widget.Toast;
 
 import com.ratajczykdev.inventoryapp.data.ImageHelper;
 import com.ratajczykdev.inventoryapp.database.Product;
-import com.ratajczykdev.inventoryapp.database.ProductListRecyclerAdapter;
 import com.ratajczykdev.inventoryapp.database.ProductViewModel;
 
 import java.util.Locale;
